% Generated by build script
\name{crates_downloads}
\alias{crates_downloads}
\title{Get download statistics for a crate}
\usage{
crates_downloads()
}
\arguments{
\item{name}{Crate name}
}
\value{tibble: date, downloads (daily download counts for last 90 days)}
\description{Get download statistics for a crate}
