% Generated by build script
\name{crates_reverse_deps}
\alias{crates_reverse_deps}
\title{Get reverse dependencies (crates that depend on this one)}
\usage{
crates_reverse_deps()
}
\arguments{
\item{name}{Crate name}
\item{per_page}{Number of results (default 20)}
}
\value{tibble: name, downloads, description}
\description{Get reverse dependencies (crates that depend on this one)}
