% Generated by build script
\name{crates_versions}
\alias{crates_versions}
\title{List all published versions of a crate}
\usage{
crates_versions()
}
\arguments{
\item{name}{Crate name (e.g. "serde", "tokio")}
}
\value{tibble: crate, version, downloads, created_at, yanked, license}
\description{List all published versions of a crate}
