% Generated by build script
\name{crates_crate}
\alias{crates_crate}
\title{Get crate details from crates.io}
\usage{
crates_crate()
}
\arguments{
\item{name}{Crate name (e.g. "serde", "tokio", "rand")}
}
\value{tibble: name, description, max_version, downloads, created_at,}
\description{Get crate details from crates.io}
